# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Theke API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/project-folders": {
      "get": { "operationId": "listProjectFolders", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Carpetas de Proyectos de la Cuenta", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectFolderListResponse" } } } } } },
      "post": { "operationId": "createProjectFolder", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectFolderInput" } } } }, "responses": { "201": { "description": "Carpeta creada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectFolderResponse" } } } } } }
    },
    "/v1/project-folders/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "patch": { "operationId": "renameProjectFolder", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectFolderInput" } } } }, "responses": { "200": { "description": "Carpeta renombrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectFolderResponse" } } } } } },
      "delete": { "operationId": "deleteProjectFolder", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Carpeta eliminada y proyectos movidos a la raíz", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" } } } } } } } } } }
    },
    "/v1/project-folders/projects": {
      "patch": { "operationId": "moveProjectsToFolder", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectFolderMoveInput" } } } }, "responses": { "200": { "description": "Proyectos movidos", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "collectionFolderId": { "type": ["string", "null"], "format": "uuid" } } } } } } } } } } }
    },
    "/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "archived"
              ],
              "default": "active"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "name": "collectionFolderId",
            "in": "query",
            "description": "UUID de carpeta o root para proyectos sin carpeta",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de proyectos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "operationId": "createProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proyecto creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/projects/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        }
      ],
      "get": {
        "operationId": "getProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Proyecto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "renameProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proyecto renombrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/projects/{id}/archive": {
      "post": {
        "operationId": "archiveProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "responses": {
          "201": {
            "description": "Proyecto archivado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/projects/{id}/restore": {
      "post": {
        "operationId": "restoreProject",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "responses": {
          "201": {
            "description": "Proyecto restaurado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/v1/projects/{projectId}/folders/{id}": {
      "patch": {
        "operationId": "updateFolder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/library-folders": {
      "get": { "operationId": "listLibraryFolders", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Carpetas de Biblioteca de la Cuenta", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LibraryFolderListResponse" } } } } } },
      "post": { "operationId": "createLibraryFolder", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LibraryFolderInput" } } } }, "responses": { "201": { "description": "Carpeta creada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LibraryFolderResponse" } } } }, "409": { "description": "Carpeta duplicada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } } } }
    },
    "/v1/library-folders/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "patch": { "operationId": "renameLibraryFolder", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LibraryFolderInput" } } } }, "responses": { "200": { "description": "Carpeta renombrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LibraryFolderResponse" } } } }, "404": { "$ref": "#/components/responses/NotFound" } } },
      "delete": { "operationId": "deleteLibraryFolder", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Carpeta eliminada y archivos movidos a la raíz", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" } } } } } } } } } }
    },
    "/v1/library-folders/resources": {
      "patch": { "operationId": "moveLibraryResources", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LibraryResourceMoveInput" } } } }, "responses": { "200": { "description": "Archivos movidos", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "libraryFolderId": { "type": ["string", "null"], "format": "uuid" } } } } } } } } } } }
    },
    "/v1/resources": {
      "get": {
        "operationId": "listResources",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "note",
                "file",
                "link"
              ]
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "archived"],
              "default": "active"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Coincidencia parcial por título, alias, descripción o contenido de nota",
            "schema": {
              "type": "string",
              "maxLength": 160
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "folderId",
            "in": "query",
            "description": "Requiere projectId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "libraryFolderId",
            "in": "query",
            "description": "UUID de carpeta o root para archivos sin carpeta",
            "schema": { "type": "string" }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": { "type": "string", "maxLength": 80 }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de recursos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/v1/resources/{id}": {
      "get": {
        "operationId": "getResource",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "ProjectId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Solicitud inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Sesión inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "No encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ProjectFolder": { "type": "object", "required": ["id", "name", "createdAt", "updatedAt"], "properties": { "id": { "type": "string", "format": "uuid" }, "name": { "type": "string" }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" } } },
      "ProjectFolderInput": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 120 } } },
      "ProjectFolderResponse": { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/ProjectFolder" } } },
      "ProjectFolderListResponse": { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectFolder" } } } },
      "ProjectFolderMoveInput": { "type": "object", "required": ["projectIds", "folderId"], "properties": { "projectIds": { "type": "array", "minItems": 1, "maxItems": 100, "items": { "type": "string", "format": "uuid" } }, "folderId": { "type": ["string", "null"], "format": "uuid" } } },
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "collectionFolderId",
          "archivedAt",
          "deletedAt",
          "purgeAfter",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "collectionFolderId": { "type": ["string", "null"], "format": "uuid" },
          "archivedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "purgeAfter": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ProjectInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "collectionFolderId": { "type": "string", "format": "uuid" }
        }
      },
      "ProjectResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Project"
          }
        }
      },
      "ProjectListResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "nextCursor"
            ],
            "properties": {
              "nextCursor": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {},
          "requestId": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
      "Folder": {
        "type": "object",
        "required": [
          "id",
          "projectId",
          "name",
          "parentFolderId",
          "archivedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "parentFolderId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "archivedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "FolderInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "parentFolderId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
      "FolderResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Folder"
          }
        }
      },
...
      "LibraryFolder": { "type": "object", "required": ["id", "name", "createdAt", "updatedAt"], "properties": { "id": { "type": "string", "format": "uuid" }, "name": { "type": "string" }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" } } },
      "LibraryFolderInput": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 120 } } },
      "LibraryFolderResponse": { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/LibraryFolder" } } },
      "LibraryFolderListResponse": { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/LibraryFolder" } } } },
      "LibraryResourceMoveInput": { "type": "object", "required": ["resourceIds", "folderId"], "properties": { "resourceIds": { "type": "array", "minItems": 1, "maxItems": 100, "items": { "type": "string", "format": "uuid" } }, "folderId": { "type": ["string", "null"], "format": "uuid" } } },
      "Resource": {
        "type": "object",
        "required": [
          "id",
          "title",
          "type",
          "origin",
          "status",
          "archivedAt",
          "updatedAt",
          "libraryFolderId",
          "aliases",
          "tags",
          "properties",
          "versionId",
          "accessibilityRequired",
          "accessibilityMissing"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "aliases": { "type": "array", "items": { "type": "string" } },
          "tags": { "type": "array", "items": { "type": "string" } },
          "properties": { "type": "object", "additionalProperties": { "oneOf": [{ "type": "string" }, { "type": "number" }, { "type": "boolean" }, { "type": "array", "items": { "type": "string" } }] } },
          "versionId": { "type": ["string", "null"], "format": "uuid" },
          "type": {
            "type": "string",
            "enum": [
              "note",
              "file",
              "link"
            ]
          },
          "mediaType": {
            "type": [
              "string",
              "null"
            ]
          },
          "byteSize": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "origin": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "archived"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "libraryFolderId": { "type": ["string", "null"], "format": "uuid" },
          "archivedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "content": {
            "type": "string"
          },
          "accessibilityText": {
            "type": [
              "string",
              "null"
            ]
          },
          "accessibilityRequired": {
            "type": "boolean"
          },
          "accessibilityMissing": {
            "type": "boolean"
          },
          "url": {
            "type": ["string", "null"],
            "format": "uri"
          },
          "previewImageUrl": {
            "type": ["string", "null"],
            "format": "uri"
          },
          "metadataStatus": {
            "type": ["string", "null"],
            "enum": ["pending", "ready", "failed", null]
          }
        }
      },
...
      "ResourceResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Resource"
          }
        }
      },
      "ResourceListResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resource"
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "nextCursor"
            ],
            "properties": {
              "nextCursor": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          }
        }
      },
...
    }
  }
}